# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Benefit Server",
    "termsOfService": "",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "13.125.61.58:9090",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/map/{latt}/{long}":{
      "get": {
        "tags": [
          "map"
        ],
        "summary": "Show all of stores or specific of stores around user without category",
        "description": "",
        "operationId": "map",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "latt",
            "in": "path",
            "description": "latitude of user",
            "required": true,
            "type": "number",
            "format": "float"
          },
          {
            "name": "long",
            "in": "path",
            "description": "longitude of user",
            "required": true,
            "type": "number",
            "format": "float"
          },
          {
            "name": "distance",
            "in": "query",
            "description": "distance around user",
            "type": "number",
            "format" : "float"
          },
          {
            "name": "X-Access-Token",
            "in": "header",
            "description": "JWT",
            "type": "apiKey"
          },
          {
            "name": "token",
            "in": "query",
            "description": "JWT",
            "type": "apiKey"
          }
        ],
        "responses": {
          "500": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/map/{latt}/{long}/{category}":{
      "get": {
        "tags": [
          "map"
        ],
        "summary": "Show all of stores or specific of stores around user with category",
        "description": "",
        "operationId": "map",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "latt",
            "in": "path",
            "description": "latitude of user",
            "required": true,
            "type": "number",
            "format": "float"
          },
          {
            "name": "long",
            "in": "path",
            "description": "longitude of user",
            "required": true,
            "type": "number",
            "format": "float"
          },
          {
            "name": "category",
            "in": "path",
            "description": "category of store",
            "required": true,
            "type": "string"
          },
          {
            "name": "distance",
            "in": "query",
            "description": "distance around user",
            "type": "number",
            "format" : "float"
          },
          {
            "name": "X-Access-Token",
            "in": "header",
            "description": "JWT",
            "type": "apiKey"
          },
          {
            "name": "token",
            "in": "query",
            "description": "JWT",
            "type": "apiKey"
          }
        ],
        "responses": {
          "500": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/api/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
...
}